Improved accuracy func integral

+++ area.c
@@ -14,13 +14,13 @@ float root(float (*f) (float), float (*g) (float),
            float a, float b, float eps1,
            float (*df) (float), float (*dg) (float));
 
-float integral(float (*f) (float), float a, float b, float eps2);
+double integral(float (*f) (float), float a, float b, float eps2);
 
 int main(int argc, char* argv[]) {
     float a = root(f1, f3, 0, 1, 0.0001, df1, df3);
     printf("%f\n", a);
     
-    float b = integral(f2, 3, 4, 0.001);
+    double b = integral(f2, 3, 4, 0.001);
     printf("%f\n", b);
    
     return 0;
