+++ integral.c
@@ -1,12 +1,12 @@
-float integral(float (*f) (float), float a, float b, float eps2);
+double integral(float (*f) (float), float a, float b, float eps2);
     
-float integral(float (*f) (float), float a, float b, float eps2) {
+double integral(float (*f) (float), float a, float b, float eps2) {
     float h = 0.00001;
     int n = (b - a) / h;
     
-    float area = 0;
+    double area = 0;
     for (int i = 0; i < n; ++i) {
-        area += f(a + (i + 0.5) * h);
+        area += (double) f(a + (i + 0.5) * h);
     }
     
     area *= h;
